move tests mutators in each folder
