+++ tests/Humbug/Test/Mutator/Boolean/BooleanAndTest.php
@@ -9,19 +9,19 @@
  * @license    https://github.com/padraic/humbug/blob/master/LICENSE New BSD License
  */
 
-namespace Humbug\Test\Mutation;
+namespace Humbug\Test\Mutator\Boolean;
 
 use Humbug\Mutator;
 
-class OperatorDecrementTest extends \PHPUnit_Framework_TestCase
+class BooleanAndTest extends \PHPUnit_Framework_TestCase
 {
 
-    public function testReturnsTokenEquivalentToIncrementOperator()
+    public function testReturnsTokenEquivalentToBooleanOr()
     {
-        $mutation = new Mutator\Increment\Decrement;
+        $mutation = new Mutator\Boolean\LogicalAnd;
         $this->assertEquals(
             array(
-                10 => array(T_INC, '++')
+                10 => array(T_BOOLEAN_OR, '||')
             ),
             $mutation->getMutation(array(), 10)
         );
