+++ tests/Humbug/Test/Mutator/Increment/OperatorDecrementTest.php
@@ -9,19 +9,19 @@
  * @license    https://github.com/padraic/humbug/blob/master/LICENSE New BSD License
  */
 
-namespace Humbug\Test\Mutation;
+namespace Humbug\Test\Mutator\Increment;
 
 use Humbug\Mutator;
 
-class OperatorSubtractionTest extends \PHPUnit_Framework_TestCase
+class OperatorDecrementTest extends \PHPUnit_Framework_TestCase
 {
 
-    public function testReturnsTokenEquivalentToAdditionOperator()
+    public function testReturnsTokenEquivalentToIncrementOperator()
     {
-        $mutation = new Mutator\Arithmetic\Subtraction;
+        $mutation = new Mutator\Increment\Decrement;
         $this->assertEquals(
             array(
-                10 => '+'
+                10 => array(T_INC, '++')
             ),
             $mutation->getMutation(array(), 10)
         );
